size_t based MIDI loader

+++ PianoFromAbove/MIDI.h
@@ -99,9 +99,9 @@ public:
     MIDIChannelEvent* AllocChannelEvent();
 
     //Parsing functions that load data into the instance
-    int ParseMIDI( const unsigned char *pcData, size_t iMaxSize );
-    int ParseTracks( const unsigned char *pcData, size_t iMaxSize );
-    int ParseEvents( const unsigned char *pcData, size_t iMaxSize );
+    size_t ParseMIDI( const unsigned char *pcData, size_t iMaxSize );
+    size_t ParseTracks( const unsigned char *pcData, size_t iMaxSize );
+    size_t ParseEvents( const unsigned char *pcData, size_t iMaxSize );
     bool IsValid() const { return ( m_vTracks.size() > 0 && m_Info.iNoteCount > 0 && m_Info.iDivision > 0 ); }
 
     void PostProcess(vector<MIDIChannelEvent*>& vChannelEvents, eventvec_t* vProgramChanges = nullptr,
@@ -157,8 +157,8 @@ public:
     ~MIDITrack( void );
 
     //Parsing functions that load data into the instance
-    size_t ParseTrack( const unsigned char *pcData, size_t iMaxSize, int iTrack );
-    size_t ParseEvents( const unsigned char *pcData, size_t iMaxSize, int iTrack );
+    size_t ParseTrack( const unsigned char *pcData, size_t iMaxSize, size_t iTrack );
+    size_t ParseEvents( const unsigned char *pcData, size_t iMaxSize, size_t iTrack );
     void clear( void );
 
     friend class MIDIPos;
